ML_Hybrid - Copy/requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
boto3==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# Text processing
spacy==3.6.0
nltk==3.8.1
ftfy==6.1.3
langdetect==1.0.9

# ML and clustering
umap-learn==0.5.3
hdbscan==0.8.29

# LangChain (latest stable versions)
langchain==0.0.350
langchain-community==0.0.10

# Vector stores
faiss-cpu==1.7.4
chromadb==0.4.15

# Advanced features
optuna==3.3.0
joblib==1.3.1

# Utilities
loguru==0.7.0
pyyaml==6.0.1
python-multipart==0.0.6

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1 